chore: Make calculateDelay a public method.
