+++ actor/src/main/scala/org/apache/pekko/pattern/BackoffOptions.scala
@@ -13,13 +13,13 @@
 
 package org.apache.pekko.pattern
 
-import scala.concurrent.duration.{ Duration, FiniteDuration }
-
 import org.apache.pekko
 import pekko.actor.{ ActorRef, OneForOneStrategy, Props, SupervisorStrategy }
 import pekko.annotation.{ DoNotInherit, InternalApi }
 import pekko.pattern.internal.{ BackoffOnRestartSupervisor, BackoffOnStopSupervisor }
 import pekko.util.JavaDurationConverters._
+
+import scala.concurrent.duration.{ Duration, FiniteDuration }
 
 /**
  * Backoff options allow to specify a number of properties for backoff supervisors.
@@ -255,6 +255,7 @@ object BackoffOpts {
       maxBackoff: java.time.Duration,
       randomFactor: Double): BackoffOnStopOptions =
     onStop(childProps, childName, minBackoff.asScala, maxBackoff.asScala, randomFactor)
+
 }
 
 /**
